feat: Add error messages to registration

+++ app/controllers/users/registrations_controller.rb
@@ -13,15 +13,16 @@ class Users::RegistrationsController < Devise::RegistrationsController
   def create
     super do |resource|
       resource.name = "@" + params[:user][:email].split('@').first
-      resource.save
-      if session[:_cart_].present?
-        cart = Cart.find(session[:_cart_])
-        cart.update(user_id: resource.id)
-        session.delete(:_cart_)
-        sign_in resource, event: :authentication
-        redirect_to carts_path and return
-      else
-        resource.create_cart()
+      if resource.save
+        if session[:_cart_].present?
+          cart = Cart.find(session[:_cart_])
+          cart.update(user_id: resource.id)
+          session.delete(:_cart_)
+          sign_in resource, event: :authentication
+          redirect_to carts_path and return
+        else
+          resource.create_cart()
+        end
       end
     end
   end
